Moved walk engine classes into utility/motion
