+++ shared/utility/motion/WalkEngine.cpp
