+++ shared/utility/motion/WalkEngine.hpp
